#  use intermediate satisfier causes in priority statistics 