This provides more information to the `DP::prioritize` call. Inspired by @x-hgg-x  investigation https://github.com/pubgrub-rs/pubgrub/pull/274#issuecomment-2504970942, it provides the number of times the package has been involved in any backtracking conflict.  Inspired by UV's  prioritization, it provides the order the package was discovered. When using `Priority = (conflict_count, Reverse<matched versions>, Reverse<discovery_order>)` Resolving all crates (including Solana) on 10 threads went from `162.45min` to `88.96min`! The slowest individual crate went from `solana-archiver-lib@1.1.0 95s` to `solana-archiver-lib@1.1.7 21s`! (Note that the runtime for any individual crate may vary run to run due to system load and scheduling.)

The new fields are hidden behind a `PackageResolutionStatistics` with getters,  to make it easier to add more in the future without additional breaking changes. Getters are used in case some value we want to return is expensive to  calculate and so we only want to calculate it if someone can read it.  I attempted to use that pattern  to avoid the hash map look up for `conflict_count`, but then `PackageResolutionStatistics` ends up generic over `P`. Which seemed unpleasant. If people have other suggestions for how to future proof this API I'm all ears!

Theoretically the package name and range  could also become getters on `PackageResolutionStatistics`. But I consider those two exponentially more important and more useful than the ones where adding now. Deserving of always being computed and getting dedicated API surface. But opinions may vary.