@konstin 2024-12-13
I found something similar for uv, where I'm also using the incompatibility (https://github.com/astral-sh/pubgrub/pull/36/commits/04005a0afc33a5372bf5c567fe1fdaf717657fc4). I also had to skip incompatibilities that were only (package A, proxy package for A) to avoid our marker package type.