@arielb1 2024-12-15
The latest solver change doesn't 100% fix the problem:
```
       Bucket:solana-sdk@1.x.y/dep:ed25519-dalek SemverPubgrub { norml: 1.18.23 | 1.18.25 | 1.18.26, pre: <E2><88><85> }  depends on Bucket:ed25519-dalek@1.x.y/default=true SemverPubgrub { norml: >=1.0.1, <1.0.2, pre: <E2><88><85> }
[2024-12-15T19:42:11Z INFO  pubgrub::internal::core] backtrack to DecisionLevel(2)
[2024-12-15T19:42:11Z INFO  pubgrub::solver] DP chose: Id::<Names>(46) = 'Bucket:solana-sdk@1.x.y/default=true' @ Some(Version { major: 1, minor: 18, patch: 22 })
[2024-12-15T19:42:11Z INFO  pubgrub::internal::partial_solution] add_decision: Id::<Names>(46) @ 1.18.22
[2024-12-15T19:42:11Z INFO  pubgrub::solver] unit_propagation: Id::<Names>(46) = 'Bucket:solana-sdk@1.x.y/default=true'
[2024-12-15T19:42:11Z INFO  pubgrub::solver] DP chose: Id::<Names>(279) = 'Bucket:ed25519-dalek@1.x.y/default' @ Some(Version { major: 1, minor: 0, patch: 0, pre: Prerelease("pre.3") })
[2024-12-15T19:42:11Z INFO  pubgrub::solver] add_decision (not first time): Id::<Names>(279) = 'Bucket:ed25519-dalek@1.x.y/default' @ 1.0.0-pre.3
[2024-12-15T19:42:11Z INFO  pubgrub::solver] unit_propagation: Id::<Names>(279) = 'Bucket:ed25519-dalek@1.x.y/default'
[2024-12-15T19:42:11Z INFO  pubgrub::solver] DP chose: Id::<Names>(280) = 'Bucket:ed25519-dalek@1.x.y/u64_backend' @ Some(Version { major: 1, minor: 0, patch: 0, pre: Prerelease("pre.3") })[2024-12-15T19:42:11Z INFO  pubgrub::solver] add_decision (not first time): Id::<Names>(280) = 'Bucket:ed25519-dalek@1.x.y/u64_backend' @ 1.0.0-pre.3
[2024-12-15T19:42:11Z INFO  pubgrub::solver] unit_propagation: Id::<Names>(280) = 'Bucket:ed25519-dalek@1.x.y/u64_backend'
[2024-12-15T19:42:11Z INFO  pubgrub::solver] DP chose: Id::<Names>(281) = 'Bucket:ed25519-dalek@1.x.y/std' @ Some(Version { major: 1, minor: 0, patch: 0, pre: Prerelease("pre.3") })
[2024-12-15T19:42:11Z INFO  pubgrub::solver] add_decision (not first time): Id::<Names>(281) = 'Bucket:ed25519-dalek@1.x.y/std' @ 1.0.0-pre.3
[2024-12-15T19:42:11Z INFO  pubgrub::solver] unit_propagation: Id::<Names>(281) = 'Bucket:ed25519-dalek@1.x.y/std'
[2024-12-15T19:42:11Z INFO  pubgrub::solver] DP chose: Id::<Names>(284) = 'Bucket:ed25519-dalek@1.x.y/rand' @ Some(Version { major: 1, minor: 0, patch: 0, pre: Prerelease("pre.3") })
[2024-12-15T19:42:11Z INFO  pubgrub::solver] add_decision (not first time): Id::<Names>(284) = 'Bucket:ed25519-dalek@1.x.y/rand' @ 1.0.0-pre.3
[2024-12-15T19:42:11Z INFO  pubgrub::solver] unit_propagation: Id::<Names>(284) = 'Bucket:ed25519-dalek@1.x.y/rand'
[2024-12-15T19:42:11Z INFO  pubgrub::solver] DP chose: Id::<Names>(286) = 'Bucket:ed25519-dalek@1.x.y/dep:rand' @ Some(Version { major: 1, minor: 0, patch: 0, pre: Prerelease("pre.3") })
[2024-12-15T19:42:11Z INFO  pubgrub::solver] add_decision (not first time): Id::<Names>(286) = 'Bucket:ed25519-dalek@1.x.y/dep:rand' @ 1.0.0-pre.3
[2024-12-15T19:42:11Z INFO  pubgrub::solver] unit_propagation: Id::<Names>(286) = 'Bucket:ed25519-dalek@1.x.y/dep:rand'
[2024-12-15T19:42:11Z INFO  pubgrub::solver] DP chose: Id::<Names>(355) = 'Bucket:solana-sdk@1.x.y/default' @ Some(Version { major: 1, minor: 18, patch: 22 })
[2024-12-15T19:42:11Z INFO  pubgrub::internal::partial_solution] add_decision: Id::<Names>(355) @ 1.18.22
[2024-12-15T19:42:11Z INFO  pubgrub::solver] unit_propagation: Id::<Names>(355) = 'Bucket:solana-sdk@1.x.y/default'
[2024-12-15T19:42:11Z INFO  pubgrub::solver] DP chose: Id::<Names>(356) = 'Bucket:solana-sdk@1.x.y/full' @ Some(Version { major: 1, minor: 18, patch: 22 })
[2024-12-15T19:42:11Z INFO  pubgrub::internal::partial_solution] add_decision: Id::<Names>(356) @ 1.18.22
[2024-12-15T19:42:11Z INFO  pubgrub::solver] unit_propagation: Id::<Names>(356) = 'Bucket:solana-sdk@1.x.y/full'
[2024-12-15T19:42:11Z INFO  pubgrub::solver] DP chose: Id::<Names>(357) = 'Bucket:solana-sdk@1.x.y/byteorder' @ Some(Version { major: 1, minor: 18, patch: 22 })
[2024-12-15T19:42:11Z INFO  pubgrub::internal::partial_solution] add_decision: Id::<Names>(357) @ 1.18.22
[2024-12-15T19:42:11Z INFO  pubgrub::solver] unit_propagation: Id::<Names>(357) = 'Bucket:solana-sdk@1.x.y/byteorder'
[2024-12-15T19:42:11Z INFO  pubgrub::solver] DP chose: Id::<Names>(358) = 'Bucket:solana-sdk@1.x.y/ed25519-dalek-bip32' @ Some(Version { major: 1, minor: 18, patch: 22 })
[2024-12-15T19:42:11Z INFO  pubgrub::internal::partial_solution] add_decision: Id::<Names>(358) @ 1.18.22
[2024-12-15T19:42:11Z INFO  pubgrub::solver] unit_propagation: Id::<Names>(358) = 'Bucket:solana-sdk@1.x.y/ed25519-dalek-bip32'
[2024-12-15T19:42:11Z INFO  pubgrub::solver] DP chose: Id::<Names>(359) = 'Bucket:solana-sdk@1.x.y/ed25519-dalek' @ Some(Version { major: 1, minor: 18, patch: 22 })
[2024-12-15T19:42:11Z INFO  pubgrub::internal::partial_solution] add_decision: Id::<Names>(359) @ 1.18.22
[2024-12-15T19:42:11Z INFO  pubgrub::solver] unit_propagation: Id::<Names>(359) = 'Bucket:solana-sdk@1.x.y/ed25519-dalek'
[2024-12-15T19:42:11Z INFO  pubgrub::solver] DP chose: Id::<Names>(373) = 'Bucket:solana-sdk@1.x.y/dep:ed25519-dalek' @ Some(Version { major: 1, minor: 18, patch: 22 })
[2024-12-15T19:42:11Z INFO  pubgrub::internal::partial_solution] not adding Id::<Names>(373) @ 1.18.22 because of its dependencies
[2024-12-15T19:42:11Z INFO  pubgrub::solver] unit_propagation: Id::<Names>(373) = 'Bucket:solana-sdk@1.x.y/dep:ed25519-dalek'
[2024-12-15T19:42:11Z INFO  pubgrub::internal::core] Start conflict resolution because incompat satisfied:
       Bucket:solana-sdk@1.x.y/dep:ed25519-dalek SemverPubgrub { norml: 1.18.22 | 1.18.23 | 1.18.25 | 1.18.26, pre: <E2><88><85> }  depends on Bucket:ed25519-dalek@1.x.y/default=true SemverPubgrub { norml: >=1.0.1, <1.0.2, pre: <E2><88><85> }
```

I would expect it to be able to prioritize this chain without "wandering":
```
solana-sdk@1.x.y/default=true -> Bucket:solana-sdk@1.x.y/default -> solana-sdk@1.x.y/full -> solana-sdk@1.x.y/ed25519-dalek -> solana-sdk@1.x.y/dep:ed25519-dalek
```

The "root cause" it finds is
```
[2024-12-15T19:44:45Z INFO  pubgrub::solver] DP chose: Id::<Names>(39) = 'Bucket:ed25519-dalek@1.x.y/default=true'
[2024-12-15T19:44:45Z INFO  pubgrub::solver] DP chose: Id::<Names>(46) = 'Bucket:solana-sdk@1.x.y/default=true'
```

Which is insufficient since it could find the rest of the packages in the chain to load them.