@konstin 2024-12-19
I'm seeing a massive speedup with solana-archiver-lib. I'm testing with 8 threads since 16 overflow with dev.

**Before** (dev, 3bef331564078fa0ece1c850518d41efe2c22b18)
```
$ cargo run -r -- -m pub --with-solana --filter solana-archiver-lib -t 8
Running in mode Pub on 8 rayon threads.
        index commit hash: 82086e46740d7a9303216bfac093e7268a95121f
        index commit time: 2024-11-30T18:18:14Z
               index size: 32
          solana in index: 32
             Pub CPU time:   257.76s ==   4.30min
           Cargo CPU time: skipped
Cargo check lock CPU time: skipped
  Pub check lock CPU time: skipped
                Wall time:    37.05s ==   0.62min
```

**After** (PR, e08db340950e3da69ba1b0e2eaa0449300b59625)
```
$ cargo run -r -- -m pub --with-solana --filter solana-archiver-lib -t 8
Running in mode Pub on 8 rayon threads.
        index commit hash: 82086e46740d7a9303216bfac093e7268a95121f
        index commit time: 2024-11-30T18:18:14Z
               index size: 32
          solana in index: 32
             Pub CPU time:   143.82s ==   2.40min
           Cargo CPU time: skipped
Cargo check lock CPU time: skipped
  Pub check lock CPU time: skipped
                Wall time:    20.17s ==   0.34min
```

It's an improvement in both measurements, thought I'm getting a very different speedup.

Counterintuitively, the should cancel count is way worse while the runtime performance improves. Is there a good non-walltime metric?

![image](https://github.com/user-attachments/assets/190c8ccd-e4ac-4303-b6d8-ec6ac5bf65bf)


Patch:

```rust
    type Priority = (u32, Reverse<usize>);

    fn prioritize(
        &self,
        package: &Names<'c>,
        range: &RcSemverPubgrub,
        package_conflicts_counts: &PackageResolutionStatistics,
    ) -> Self::Priority {
        let matches_count = match package {
            Names::Links(_name) => {
                // PubGrub automatically handles when any requirement has no overlap. So this is only deciding a importance of picking the version:
                //
                // - If it only matches one thing, then adding the decision with no additional dependencies makes no difference.
                // - If it can match more than one thing, and it is entirely equivalent to picking the packages directly which would make more sense to the users.
                //
                // So only rubberstamp links attributes when all other decisions are made, by setting the priority as low as it will go.
                usize::MAX
            }

            Names::Wide(_, req, _, _) => self.count_wide_matches(range, &package.crate_(), req),
            Names::WideFeatures(_, req, _, _, _) | Names::WideDefaultFeatures(_, req, _, _) => self
                .count_wide_matches(range, &package.crate_(), req)
                .saturating_add(1),

            Names::Bucket(_, _, _) => self.count_matches(range, &package.crate_()),
            Names::BucketFeatures(_, _, _) | Names::BucketDefaultFeatures(_, _) => self
                .count_matches(range, &package.crate_())
                .saturating_add(1),
        };
        (
            package_conflicts_counts.conflict_count(),
            Reverse(matches_count),
        )
    }
```

(I'll still have to give the updated PR a proper read)